RAFT Fixed notebook 3 (#7)
* 🎨 Isolated headers az creadentials compute cell

Removed output cells from the deployment notebook to enhance clarity and prevent clutter during execution. This change ensures that the notebook remains clean and focused on the code and its functionality.

* 🎨 Update next step links in 1_gen.ipynb

Removed the reference to the Azure MaaS fine-tuning notebook in the next steps section for clarity.

* 🎨 Clear execution outputs in fine-tuning notebook

Removed execution outputs and set execution counts to null in the fine-tuning notebook to enhance clarity and prevent clutter during model training.

* Update evaluation notebook: reset execution counts, fix environment variable name, and enhance output messages. Upgrade Python version to 3.12.8.